feat(veu):Adds agency VEU project activation endpoint
Adds an endpoint to activate VEU projects for agency properties.

This change allows administrators to trigger the activation of
VEU (Verified Energy Upgrade) projects for agencies, initiating
the process of creating related data and configurations.

The endpoint returns a summary of the inserted records,
enhancing transparency and providing feedback on the operation's
success.